refactor: replace sologame/triogame entities with litegame
these all share common base classes, and solo records will ideally
be migrated to lite records once all of the other components are

TODO: uncomment/migrate lite entities
