+++ server/src/entity/LitePlayerDecision.ts
@@ -1,43 +1,36 @@
-import {
-    Entity,
-    PrimaryGeneratedColumn,
-    Column,
-    ManyToOne,
-    JoinColumn,
-} from "typeorm";
-import { MultiplayerGameRound } from "./LiteGameRound";
-import { MultiplayerPlayer } from "./LitePlayer";
-import { MultiplayerMarsEventDeck } from "./MultiplayerMarsEventDeck";
-  
+import { Entity, PrimaryGeneratedColumn, Column, ManyToOne, JoinColumn } from "typeorm";
+// import { LiteGameRound } from "./LiteGameRound";
+// import { LitePlayer } from "./LitePlayer";
+// import { LiteMarsEventDeck } from "./LiteMarsEventDeck";
+
 export abstract class BaseLitePlayerDecision {
-    @PrimaryGeneratedColumn()
-    id!: number;
-  
-    @Column()
-    systemHealthInvestment!: number;
-  
-    @Column()
-    pointsInvestment!: number;
+  @PrimaryGeneratedColumn()
+  id!: number;
+
+  @Column()
+  systemHealthInvestment!: number;
+
+  @Column()
+  pointsInvestment!: number;
 }
-  
+
 @Entity()
 export class SoloPlayerDecision extends BaseLitePlayerDecision {}
-  
-@Entity()
-export class MultiplayerPlayerDecision extends BaseLitePlayerDecision {
-    @ManyToOne(() => MultiplayerGameRound, round => round.decisions)
-    round!: MultiplayerGameRound;
-  
-    @ManyToOne(() => MultiplayerPlayer, player => player.decisions)
-    player!: MultiplayerPlayer;
-  
-    @Column()
-    vote!: string;
-  
-    @ManyToOne(() => MultiplayerPlayer, { nullable: true })
-    votedPlayer!: MultiplayerPlayer;
-  
-    @ManyToOne(() => MultiplayerMarsEventDeck, { nullable: true })
-    event!: MultiplayerMarsEventDeck;
-}
-  
+
+// @Entity()
+// export class LitePlayerDecision extends BaseLitePlayerDecision {
+//   @ManyToOne(() => LiteGameRound, round => round.decisions)
+//   round!: LiteGameRound;
+
+//   @ManyToOne(() => LitePlayer, player => player.decisions)
+//   player!: LitePlayer;
+
+//   @Column()
+//   vote!: string;
+
+//   @ManyToOne(() => LitePlayer, { nullable: true })
+//   votedPlayer!: LitePlayer;
+
+//   @ManyToOne(() => LiteMarsEventDeck, { nullable: true })
+//   event!: LiteMarsEventDeck;
+// }
