--- server/src/entity/SoloHighScore.ts
@@ -1,46 +1,0 @@
-import {
-  Column,
-  Entity,
-  OneToOne,
-  PrimaryGeneratedColumn,
-  UpdateDateColumn,
-  CreateDateColumn,
-  ManyToOne,
-  JoinColumn,
-} from "typeorm";
-import { User } from "./User";
-import { SoloGame } from "./SoloGame";
-
-@Entity()
-export class SoloHighScore {
-  @PrimaryGeneratedColumn()
-  id!: number;
-
-  @ManyToOne(type => User, { nullable: false })
-  user!: User;
-
-  @Column()
-  userId!: number;
-
-  @Column("float")
-  pointsPerRound!: number;
-
-  @Column()
-  points!: number;
-
-  @Column()
-  maxRound!: number;
-
-  @OneToOne(type => SoloGame, { nullable: true })
-  @JoinColumn()
-  game!: SoloGame;
-
-  @Column({ nullable: true })
-  gameId!: number;
-
-  @CreateDateColumn()
-  dateCreated!: Date;
-
-  @UpdateDateColumn()
-  lastModified!: Date;
-}
