--- server/src/entity/SoloMarsEventCard.ts
@@ -1,44 +1,0 @@
-import { Column, Entity, PrimaryGeneratedColumn } from "typeorm";
-import { LiteGameType } from "@port-of-mars/shared/lite";
-
-@Entity()
-export class SoloMarsEventCard {
-  @PrimaryGeneratedColumn()
-  id!: number;
-
-  @Column({ default: "freeplay" })
-  gameType!: LiteGameType;
-
-  @Column()
-  codeName!: string;
-
-  @Column({ default: "" })
-  displayName!: string;
-
-  @Column({ default: "" })
-  flavorText!: string;
-
-  @Column()
-  effect!: string;
-
-  @Column({ default: 1 })
-  drawMin!: number;
-
-  @Column({ default: 1 })
-  drawMax!: number;
-
-  @Column({ default: 0 })
-  rollMin!: number;
-
-  @Column({ default: 0 })
-  rollMax!: number;
-
-  @Column({ default: 0 })
-  systemHealthMultiplier!: number;
-
-  @Column({ default: 0 })
-  pointsMultiplier!: number;
-
-  @Column({ default: 0 })
-  resourcesMultiplier!: number;
-}
