--- server/src/entity/SoloPlayerDecision.ts
@@ -1,13 +1,0 @@
-import { Entity, PrimaryGeneratedColumn, Column } from "typeorm";
-
-@Entity()
-export class SoloPlayerDecision {
-  @PrimaryGeneratedColumn()
-  id!: number;
-
-  @Column()
-  systemHealthInvestment!: number;
-
-  @Column()
-  pointsInvestment!: number;
-}
