+++ server/src/entity/index.ts
@@ -1,3 +1,5 @@
+import exp from "constants";
+
 export * from "./ChatReport";
 export * from "./Game";
 export * from "./GameEvent";
@@ -12,13 +14,14 @@ export * from "./Treatment";
 export * from "./TournamentRound";
 export * from "./TournamentRoundInvite";
 export * from "./User";
-export * from "./SoloGame";
-export * from "./SoloGameRound";
-export * from "./SoloGameTreatment";
-export * from "./SoloMarsEventCard";
-export * from "./SoloMarsEventDeck";
-export * from "./SoloMarsEventDeckCard";
-export * from "./SoloPlayer";
-export * from "./SoloPlayerDecision";
+export * from "./LiteGame";
+export * from "./LiteGameRound";
+export * from "./LiteGameTreatment";
+export * from "./LiteHighScore";
+export * from "./LiteMarsEventCard";
+export * from "./LiteMarsEventDeck";
+export * from "./LiteMarsEventDeckCard";
+export * from "./LitePlayer";
+export * from "./LitePlayerDecision";
 export * from "./ProlificStudy";
 export * from "./ProlificStudyParticipant";
